requirements.txt
google-generativeai>=0.8.0
google-ai-generativelanguage>=0.6.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
protobuf>=4.25.0
requests>=2.31.0
reportlab>=4.0.0
beautifulsoup4>=4.13.0
weasyprint>=61.0
pdfkit>=1.0.0

# API Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0